chore: move /parse_csv to /csv
